Extract useOrigin hook
