+++ src/components/safe-messages/InfoBox/index.tsx
@@ -23,9 +23,7 @@ const InfoBox = ({
           <Typography variant="subtitle1" fontWeight="bold">
             {title}
           </Typography>
-          <Typography variant="body2" display="inline">
-            {message}{' '}
-          </Typography>
+          <Typography variant="body2">{message}</Typography>
         </div>
       </div>
       <Divider className={css.divider} />
