+++ src/components/transactions/TxShareLink/index.tsx
@@ -1,5 +1,4 @@
 import type { ReactElement } from 'react'
-import { useEffect, useState } from 'react'
 import { IconButton, Link, SvgIcon } from '@mui/material'
 import ShareIcon from '@/public/images/common/share.svg'
 import { AppRoutes } from '@/config/routes'
@@ -8,17 +7,7 @@ import Track from '@/components/common/Track'
 import { TX_LIST_EVENTS } from '@/services/analytics/events/txList'
 import React from 'react'
 import CopyTooltip from '@/components/common/CopyTooltip'
-
-const useOrigin = () => {
-  const [origin, setOrigin] = useState('')
-
-  useEffect(() => {
-    if (typeof location !== 'undefined') {
-      setOrigin(location.origin)
-    }
-  }, [])
-  return origin
-}
+import useOrigin from '@/hooks/useOrigin'
 
 const TxShareLink = ({ id }: { id: string }): ReactElement => {
   const router = useRouter()
